+++ core-jobs/db/migrate/20221006214505_create_que_schema.rb
@@ -5,17 +5,15 @@ class CreateQueSchema < ActiveRecord::Migration[6.0]
     # Whenever you use Que in a migration, always specify the version you're
     # migrating to. If you're unsure what the current version is, check the
     # changelog.
+    
+    return if Que.db_version >= 8 # Que is already installed at the highest version
 
     if Que.db_version.zero? # Que is not installed here so install the highest version
       Que.migrate!(version: 8)
       return
     end
 
-    raise 'Que version must be updated to v5 before core-jobs can be installed' if Que.db_version < 5
-
-    Que.migrate!(version: 6) if Que.db_version == 5
-    Que.migrate!(version: 7) if Que.db_version == 6
-    Que.migrate!(version: 8) if Que.db_version == 7
+    (Que.db_version + 1..8).each { |v| Que.migrate!(version: v) } # catch que up to the highest version
   end
 
   def down