+++ core-jobs/lib/core/jobs.rb
@@ -12,7 +12,9 @@ require 'que/active_record/model'
 
 require 'core/jobs/puma/plugin/que'
 require 'core/jobs/railtie' if defined?(Rails::Railtie)
+require 'core/jobs/application_job'
 require 'core/jobs/que_job_ext'
+require 'core/jobs/que_job'
 require 'core/jobs/job'
 require 'core/jobs/locker'
 
@@ -30,9 +32,8 @@ require 'core/jobs/graphql/job/report/queries'
 require 'core/jobs/graphql/job/report/subscriptions'
 
 if defined?(Wisper)
-  require 'core/jobs/publishers/que_state'
-  require 'core/jobs/subscribers/que_state'
-  require 'core/jobs/listener'
+  require 'core/jobs/pub_sub'
+  require 'core/jobs/listeners'
 end
 
 # Dir["core/jobs/graphql/**/*.rb"].each { |file| require file }
@@ -45,9 +46,3 @@ module Core
     class WrongActiveJobQueueAdapter < Core::Jobs::Error; end
   end
 end
-
-# if Rails.env.development?
-#   Rails.application.console do
-#     puts "Custom message here"
-#   end
-# end